+++ CHANGELOG.md
@@ -13,6 +13,7 @@ master
 * Updated symfony/var-dumper to 6.4
 * Added empty() to the list of banned functions
 * Added PHP 8.4 to the CI run
+* Added support for nested namespaces for tests in `BannedUseTestRule`
 
 v3.0.0
 ------